Started working on the parser for the markdown checklist.md.
Once these can be parsed simple enough, I can than form a hierarhcy of entries like the following:

A CheckList_Entry which has two core properties a name and a list of info. The name comes from what is currently a header, and the info nodes start with the equals line and if there are volume entries before the next header entry, append those to the list of article infos, so that it can all be collected.

The end goal is to reduce the total amount of work I actually need to do in gathering the info. Once I can reliably reproduce the info from this, generating a new spreadsheet will be easy, as will any future endevours I hope.