The pieces have been linked! and it kinda works so far. the first few entries of BGU at least can be found (should probably look into why its skipping over Actenstücke and jumping right to BGU, but thats one of the many issues which need to be worked out. Another major issue is continuing to both refine the data set I'm using, but also refine my code for analyzing it, and refine my code to process it.
By this I really mean expanding advanced search. Currently its patchwork, but it needs to be refined for sure for BGU XVII, which is 16619, but when you search for the BGU entry the wrong BGU is pulled up.

Likely expanding this could be done by taking the , ed. element which should be stored in the full texts and splitting on that for a bit.

Might ask ChatGPT or similar program to help write some regex to break apart the text into its composite parts, so that I can process it that way more easily.

Basically for the next advanced search, its taking one of those parts, the author or the publication year (I realize that this might be easier to extract with regex, maybe assuming any combiniation of 18XX, 19XX, 20XX + a period at the end. Maybe split by Nos. first, and then take everything before the nos, and find the date in that bit?)

Anyway once we have the author, date, or both, supply those to the search system and have it use them to help deduce the correct result.