stuff I forgot to include in the previous commit.
basically current issue if BGU V is failing because the actual answer doesn't match well enough, but somehow also the correct result isn't getting set so its failing or something in the print. I'm gonig to quickly try and fix that second issue, but let the larger ranking failure remain for now. The ranking can live if we can just turn it into a thing where the user either has to confirm or pick from a set of rankings which one is correct to enter into the spreadsheet. Then it should be much easier to do I hope.